Refactor: MVC 패턴 적용
- model : 유효값 검사, 정답 검사
- view : 입,출력
- controller : BridgeGameControl.js
- utils : 상수값
